production: properly reset machine in any case
We encountered an issue that the delivery station was not behaving
properly after breaking once for the rest of the game.
Symptom: Machine is Prepared but does not switch to PROCESSING
Issue: Machine timeout due to processing does not clean up the set task
which prevents the subsequent prepared attempts to get stuck and never
move to processing.

Fix: actually properly reset a machines logical state when recovering
from Broken, such that there is no distinction beteen time out on
PREPARED and on PROCESSING (because timeout on PROCESSING.

Actually there is a race between timeout on processsing and on prepared
that makes this issue non-deterministics (as the timeout-on-prepared
rule has an exception for the DS specifically)......