rcll: load game paramters from config after pre-game
This is a crude way to enable paramterization from the frontend before
the setup phase is hit.
It is not flexible in the sense that once the pre-game is over, no
parameterization can be done anymore, but for now it should be a solid
middle-ground until a nice and clean way of reconfiguring arbitrary yaml
parameters at runtime is designed.