rcll: do not break machines when prepared in READY-AT-OUTPUT state
This causes constant pain for teams that send prepare messages without
properly waiting for feedback.

Note that teams still need to be careful at the DS to not accidentally
send deliveries twice (causing the first to go through and the second to
eventually timeout the machine), but overall it should provide a much
smoother experience this way.