rcll: obsolete wait-for-product-since, rework proc-start
The two fields were kinda similar and not consistently used properly.

wait-for-product-since was primarily used between prepared and
processing, but stations such as the storage or base station do not
start with retrieving a product from the input, hence the usage was
confusing.

proc-start used to kinda only track the time after processing, in
particularly, it regularly did not distinguish between the actual
production operation (mount ring/cap) and the following belt movement.

As we want to properly each "atomic" operation separately, this makes no
sense. Hence, we should rather update it after each machine state.

THe policy now is to update it on PREPARED, PROCESSING, PROCESSED and
READY-AT-OUTPUT and to unify the associated referee reminders
independent on the actual state.